server-app/requirements.txt
flask
opencv-python-headless
torch
Pillow
numpy
ultralytics